rearranged some tests
