+++ e2e_tests/payment_test.py
@@ -1,7 +1,6 @@
 import os
 import unittest
 from unit import Unit
-from unit.models.codecs import DtoDecoder
 from unit.models.payment import *
 from e2e_tests.account_test import create_deposit_account
 
@@ -75,63 +74,3 @@ def test_update_book_payment():
     request = PatchBookPaymentRequest(payment_id, tags)
     response = client.payments.update(request)
     assert response.data.type == "bookPayment"
-
-def test_ach_received_payment_dto():
-    ach_received_payment_response_json_api = {
-          "type": "achReceivedPayment",
-          "id": "1337",
-          "attributes": {
-            "createdAt": "2022-02-01T12:03:14.406Z",
-            "status": "Completed",
-            "wasAdvanced": True,
-            "amount": 100000,
-            "completionDate": "2022-01-23",
-            "companyName": "Uber",
-            "counterpartyRoutingNumber": "051402372",
-            "description": "Sandbox Transaction",
-            "traceNumber": "123456789123456",
-            "secCode": "PPD",
-            "tags": {}
-          },
-          "relationships": {
-            "account": {
-              "data": {
-                "type": "account",
-                "id": "163555"
-              }
-            },
-            "customer": {
-              "data": {
-                "type": "customer",
-                "id": "129522"
-              }
-            },
-            "receivePaymentTransaction": {
-              "data": {
-                "type": "transaction",
-                "id": "101"
-              }
-            },
-            "paymentAdvanceTransaction": {
-              "data": {
-                "type": "transaction",
-                "id": "202"
-              }
-            },
-            "repayPaymentAdvanceTransaction": {
-              "data": {
-                "type": "transaction",
-                "id": "890"
-              }
-            }
-          }
-        }
-
-    _id = ach_received_payment_response_json_api["id"]
-    attributes = ach_received_payment_response_json_api["attributes"]
-    _type = ach_received_payment_response_json_api["type"]
-
-    payment = DtoDecoder.decode(ach_received_payment_response_json_api)
-
-    assert payment.id == _id
-    assert str(payment.attributes["completionDate"]) == attributes["completionDate"]
